+++ pkg/ui/init.go
@@ -1,0 +1,10 @@
+package ui
+
+import (
+	tea "github.com/charmbracelet/bubbletea"
+)
+
+
+func (m Model) Init() tea.Cmd {
+	return tea.EnterAltScreen
+}
