+++ pkg/ui/model.go
@@ -1,0 +1,63 @@
+package ui
+
+import (
+	"github.com/charmbracelet/bubbles/key"
+	"github.com/charmbracelet/bubbles/list"
+	"github.com/charmbracelet/bubbles/textinput"
+)
+
+type Model struct {
+	menu          list.Model
+	display       textinput.Model
+	list          list.Model
+	itemGenerator *randomItemGenerator
+	keys          *listKeyMap
+	delegateKeys  *delegateKeyMap
+}
+
+type item struct {
+	title       string
+	description string
+}
+
+func (i item) Title() string       { return i.title }
+func (i item) Description() string { return i.description }
+func (i item) FilterValue() string { return i.title }
+
+func NewModel() Model {
+	var (
+		itemGenerator randomItemGenerator
+		delegateKeys  = newDelegateKeyMap()
+		listKeys      = newListKeyMap()
+	)
+
+	// Make initial list of items
+	const numItems = 24
+	items := make([]list.Item, numItems)
+	for i := 0; i < numItems; i++ {
+		items[i] = itemGenerator.next()
+	}
+
+	// Setup list
+	delegate := newItemDelegate(delegateKeys)
+	groceryList := list.New(items, delegate, 0, 0)
+	groceryList.Title = "Groceries"
+	groceryList.Styles.Title = titleStyle
+	groceryList.AdditionalFullHelpKeys = func() []key.Binding {
+		return []key.Binding{
+			listKeys.toggleSpinner,
+			listKeys.insertItem,
+			listKeys.toggleTitleBar,
+			listKeys.toggleStatusBar,
+			listKeys.togglePagination,
+			listKeys.toggleHelpMenu,
+		}
+	}
+
+	return Model{
+		list:          groceryList,
+		keys:          listKeys,
+		delegateKeys:  delegateKeys,
+		itemGenerator: &itemGenerator,
+	}
+}
